+++ src/Controllers/GestorGenerico.java
@@ -9,7 +9,8 @@ import java.util.Map;
 
 public class GestorGenerico<T extends IJson> {
     //Lista.
-    protected HashMap<Integer,T> lista;
+    // SE USA LONG PURA Y EXCLUSIVAMENTE POR QUE LA CLAVE DE LOS PROVEEDORES ES EL CUIL (NUMERO DE 11 DIGITOS).
+    protected HashMap<Long,T> lista;
 
     //Método constructor.
     public GestorGenerico(){
@@ -19,7 +20,7 @@ public class GestorGenerico<T extends IJson> {
 
     //Métodos propios.
         //Método agregar a lista.
-    public Boolean agregar(Integer clave,T elemento) throws ElementoRepetidoException {
+    public Boolean agregar(Long clave,T elemento) throws ElementoRepetidoException {
         if(elemento == null) throw new NullPointerException();
         if(lista.containsKey(clave)) throw new ElementoRepetidoException();
         lista.put(clave, elemento);
@@ -27,14 +28,14 @@ public class GestorGenerico<T extends IJson> {
     }
 
         //Método eliminar a lista.
-    public Boolean eliminar(Integer clave) throws ElementoNoEncontradoException {
+    public Boolean eliminar(Long clave) throws ElementoNoEncontradoException {
         if(!lista.containsKey(clave)) throw new ElementoNoEncontradoException();
         lista.remove(clave);
         return true;
     }
 
         //Método buscar en la lista.
-    public T buscarPorId(Integer idBuscar) throws ElementoNoEncontradoException {
+    public T buscarPorId(Long idBuscar) throws ElementoNoEncontradoException {
         if(lista.containsKey(idBuscar)){
             return lista.get(idBuscar);
         }
@@ -44,19 +45,19 @@ public class GestorGenerico<T extends IJson> {
         //Mostrar lista.
     public String mostrarLista(){
         StringBuilder sb = new StringBuilder("- Lista -\n");
-        for(Map.Entry<Integer, T> entry : lista.entrySet()){
+        for(Map.Entry<Long, T> entry : lista.entrySet()){
             sb.append(entry.getValue().toString()).append("\n");
         }
         return sb.toString();
     }
 
-    public HashMap<Integer,T> getMap(){
+    public HashMap<Long,T> getMap(){
         return lista;
     }
 
-    public static <T> HashSet exportarASet(HashMap<Integer,T> elementos){
+    public static <T> HashSet exportarASet(HashMap<Long,T> elementos){
         HashSet<T> set = new HashSet<>();
-        for(Map.Entry<Integer,T> entry : elementos.entrySet()){
+        for(Map.Entry<Long,T> entry : elementos.entrySet()){
             set.add(entry.getValue());
         }
         return set;
@@ -65,7 +66,7 @@ public class GestorGenerico<T extends IJson> {
         //toJson.
     public JSONArray toJson(){
         JSONArray listaJSON = new JSONArray();
-        for(Map.Entry<Integer, T> entry : lista.entrySet()){
+        for(Map.Entry<Long, T> entry : lista.entrySet()){
             listaJSON.put(entry.getValue().toJson());
         }
         return listaJSON;
